изменил сущности

+++ app/src/main/java/com/mironov/coursework/data/network/model/message/ReactionDto.kt
@@ -7,5 +7,5 @@ import kotlinx.serialization.Serializable
 data class ReactionDto(
     @SerialName("emoji_code") val emojiCode: String,
     @SerialName("emoji_name") val emojiName: String,
-    @SerialName("user_id") val userId: Long,
+    @SerialName("user_id") val userId: Int,
 )
