+++ app/src/main/java/com/mironov/coursework/data/network/user/UserDto.kt
@@ -8,5 +8,5 @@ data class UserDto(
     @SerialName("avatar_url") val avatarUrl: String?,
     @SerialName("full_name") val userName: String,
     @SerialName("email") val email: String,
-    @SerialName("user_id") val userId: Long
+    @SerialName("user_id") val userId: Int
 )
